+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 **Version 1.6.3**
 
-- SimpleNaaccrValidatable will convert the sex value to SSF25 when the SSF25 discriminator is missing for Peritoneum and Peritoneum Female Gen so that the CS schema will be found.
+- SimpleNaaccrLinesValidatable will convert the sex value to SSF25 when the SSF25 discriminator is missing for Peritoneum and Peritoneum Female Gen so that the CS schema will be found.
 
 **Version 1.6.2**